requirements.txt
attrs==20.3.0
autobahn==21.2.2
Automat==20.2.0
backoff==1.10.0
CacheControl==0.12.6
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.6
dateparser==1.0.0
firebase-admin==4.5.2
google-api-core==1.26.0
google-api-python-client==1.12.8
google-auth==1.27.0
google-auth-httplib2==0.0.4
google-cloud-core==1.6.0
google-cloud-firestore==2.0.2
google-cloud-storage==1.36.1
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
grpcio==1.36.0
httplib2==0.19.0
hyperlink==21.0.0
idna==2.10
incremental==17.5.0
kucoin-python==1.0.5
msgpack==1.0.2
packaging==20.9
proto-plus==1.14.2
protobuf==3.15.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
ratelimit==2.2.1
regex==2020.11.13
requests==2.25.1
rsa==4.7.2
service-identity==18.1.0
six==1.15.0
Twisted==21.2.0
txaio==21.2.1
tzlocal==2.1
ujson==4.0.2
uritemplate==3.0.1
urllib3==1.26.3
websockets==8.1
zope.interface==5.2.0
